Split up background.js and fix lint issues (eslint wasn't getting run somehow?)
