+++ src/bootstrap.js
@@ -2,6 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+/* eslint-disable no-unused-vars */
+
 const {utils: Cu} = Components;
 
 function startup({webExtension}, reason) {
