+++ src/webextension/manage/components/entryList.js
@@ -2,7 +2,6 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { Localized } from "fluent-react/compat";
 import PropTypes from "prop-types";
 import React from "react";
 
@@ -28,5 +27,6 @@ EntryList.propTypes = {
       site: PropTypes.string.isRequired,
     }).isRequired
   ).isRequired,
+  selected: PropTypes.number,
   onEntryClick: PropTypes.func.isRequired,
 };
